EDMS/contracts/serializers.py
from contracts.models import Contract
from rest_framework import serializers


class ContractSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contract
        fields = "__all__"
