pyproject.toml
[tool.poetry]
name = "EDMS"
version = "0.1.0"
description = ""
authors = ["Krzychu <czernicki.krz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-environ = "^0.11.2"
requests = "^2.31.0"
django-extensions = "^3.2.3"
django-filter = "^23.5"
humanize = "^4.9.0"
pre-commit = "^3.6.0"
psycopg = "^3.1.17"
psycopg2-binary = "^2.9.9"
django-cleanup = "^8.0.0"
pillow = "^10.2.0"
django-session-timeout = "^0.1.0"
celery = "^5.3.6"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
django-plotly-dash = "^2.3.0"
dpd-static-support = "^0.0.5"
channels = "^4.0.0"
daphne = "^4.1.0"
channels-redis = "^4.2.0"
djangorestframework = "^3.15.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
